% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{breed}
\alias{breed}
\title{Simulate the progeny}
\usage{
breed(pair, litter)
}
\arguments{
\item{pair}{A vector of alleles, of length 2.
Alleles must be represented as 0, 1 or 2.}

\item{litter}{The litter size from the breeding pair.}
}
\value{
The alleles of progeny, also using the 0, 1 or 2 representation
}
\description{
Simulates the progeny for a given breeding pair, to a specified litter size
}
\examples{
breed(c(1, 1))

# Sampling a population
pop <-matrix(sample(c(0, 1, 2), 100, TRUE), ncol = 2)
progeny <- apply(pop, 1, breed)
}

