% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{breedInPairs}
\alias{breedInPairs}
\title{Simulate the progeny}
\usage{
breedInPairs(pairs, litter)
}
\arguments{
\item{pairs}{A two-column matrix of alleles.
Alleles must be represented as 0, 1 or 2.}

\item{litter}{The litter size from the breeding pair.}
}
\value{
The alleles of progeny, also using the 0, 1 or 2 representation. Returned as a vector
}
\description{
Simulates the progeny for a set of breeding pairs, to a specified litter size
}
\examples{
breed(c(1, 1), litter = 6)

# Sampling a population
pop <-matrix(sample(c(0, 1, 2), 10, TRUE), ncol = 2)
apply(pop, 1, breed, litter = 6)
breedInPairs(pop, litter = 6)
}

