% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenic_drift.R
\name{calculate_deltaV_matrix}
\alias{calculate_deltaV_matrix}
\title{Calculate binding avidity adaptation matrix}
\usage{
calculate_deltaV_matrix(viruspars, maxV = 3, maxK = 25, step = 1)
}
\arguments{
\item{viruspars}{the vector of virus pars as taken by \link{\code{run_simulation}}}

\item{maxV}{the maximum binding avidity to calculate for}

\item{maxK}{the maximum achievable antibody titre in which the virus will have to adapt}

\item{step}{over how much time does this adaptation take place? usually 1 day}
}
\value{
a numeric matrix
}
\description{
Calculates the gradient of binding avidity adaptation for different levels of host immunity.
This is useful because we need to solve some ODEs for each combination of host immunity and
each possible binding avidity value ie. precomputation
}

