% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countKs}
\alias{countKs}
\title{Counts Ks}
\usage{
countKs(ks, N)
}
\arguments{
\item{ks}{the vector of K values}

\item{N}{the maximum K value to count}
}
\value{
a vector of length N with K counts
}
\description{
Counts the number of individuals in the host population with each K value from 0 to N
}
