% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftSim-Package.R
\docType{package}
\name{driftSim}
\alias{driftSim}
\alias{driftSim-package}
\title{driftSim - S Pckage for Simulating Genetic Drift}
\description{
driftSim - S Pckage for Simulating Genetic Drift
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

