% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleParameters}
\alias{exampleParameters}
\title{Example parameters}
\format{A list with 4 elements:
\describe{
\item{\strong{hostpars}, the vector of host parameters:
\itemize{
\item{S0: Starting susceptible population size}
\item{I0: Initial infecteds}
\item{R0: Initial recovered}
\item{c: Contact rate per day (ie. each person makes on average 0.7 contacts per day)}
\item{mu: Birth and death rate per day}
\item{w: Immunity waning rate}
\item{g: Recovery rate}
\item{iniBind: Initial binding avidity of seed virus}
\item{meanBoost: Average (of poisson) antibody boosting upon recovery}
\item{iniDist: Initial antigenic distance of seed virus to "root"}
\item{saveFreq: How often the distribution of host immunity values are saved}
\item{maxTitre: Maximum attainable titre for hosts}
}}
\item{\strong{viruspars}, the vector of virus parameters:
\itemize{
\item{p: parameter to control degree by which changes in binding avidity affect probability of escape from immune response}
\item{r: parameter to control degree by which previous exposure reduce probability of immune escape}
\item{q: parameter to control the shape of the relationship between binding avidity and immune escape (shift on the x-axis)}
\item{a: controls rate of changes of relationship between probability of successful replication and change in binding avidity}
\item{b: parameter to control the shape of the relationship between probability of successful replication and changes in binding avidity}
\item{n: number of offspring per virus replication event}
\item{v: number of virions initially transmitted}
\item{probMut: Probability of a mutation in antigenic properties in a given infection}
\item{expDist: Average of the exponential distribution from which the size of antigenic mutatations is drawn}
\item{kc: Rate of binding avidity adaptation}
\item{VtoD: Proportion mediating how much antigenic change occurs given a unit of binding avidity change}
}}
\item{\strong{flags}, the vector of flags for saving output options:
\itemize{
\item{SIR_flag: Flag to save SIR dynamics}
\item{voutput_phylogenetic_flag: Flag to save virus information for phylogenetic tree analysis (ie. virus parent information)}
\item{voutput_pairwise_flag: Flag to save pairwise distance matrix}
\item{time_flag: Flag to record time taken for simulation}
\item{save_state: Flag to save the final state of the simulation}
\item{input_flag_generated: Flag to use specified file as input for simulation}
\item{input_flag_saved: Flag to use specified file as input for simulation}
\item{save_k: Flag to save the distribution of hosts in each K value at regular intervals}
}}
\item{\strong{deltaVMat}, the numeric matrix of binding avidity vs. host immunity gradients for working out within-host binding avidity adaptation}
}}
\usage{
attach(exampleParameters)
}
\description{
A list of parameters to use for the simulation as an example.
}
\seealso{
\code{\link{calculate_deltaV_matrix}}, \code{\link{run_simulation}}
}
\keyword{datasets}

