% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenic_drift.R
\name{generateHostKDist}
\alias{generateHostKDist}
\title{Generate distribution of host immunity from file}
\usage{
generateHostKDist(hostFile, N)
}
\arguments{
\item{hostFile}{the file name to be read in (a csv file)}

\item{N}{number of samples to generate}
}
\value{
a vector of random K values of size N
}
\description{
Given a file which contains a column called "hostK", uses this column to generate a host K distribution as in \code{\link{generateHostKDist_2}}
}
\seealso{
\code{\link{generateHostKDist_2}}
}
