% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateKSamples}
\alias{generateKSamples}
\title{Generate N samples of immune values (K)}
\usage{
generateKSamples(cumSumK, N)
}
\arguments{
\item{cumSumK}{the cumulative distribution of K frequencies}

\item{N}{the number samples to draw}
}
\value{
A random sample of N K values from cumSumK
}
\description{
Generate N samples of immune values (K)
}
\seealso{
\code{\link{generateStartK}}
}

