% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateStartK}
\alias{generateStartK}
\title{Draws a single starting immune value from a cumulative vector distribution}
\usage{
generateStartK(dist)
}
\arguments{
\item{dist}{the vector of K values to draw from.}
}
\value{
A single K value from the given distribution
}
\description{
Draws a single starting immune value from a cumulative vector distribution
}
\seealso{
\code{\link{generateKSamples}}
}
