% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenic_drift.R
\name{plot_SIR}
\alias{plot_SIR}
\title{SIR dynamics plot}
\usage{
plot_SIR(dat)
}
\arguments{
\item{dat}{the two dimensional matrix of S, I and R data}
}
\value{
a ggplot2 object plot
}
\description{
Plots SIR dynamics for a given matrix of SIR time series data
}

