% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{run_simulation}
\alias{run_simulation}
\title{Main simulation interface}
\usage{
run_simulation(
  flags = c(1, 0, 0, 0),
  hostpars = c(90000, 100, 0, 1.5, 1/(40 * 365), 1/25, 0, 0),
  vlPars,
  infectiousnessPars,
  crossImmunity,
  start = 0,
  end = 100,
  output_files = c("SIR.csv", "voutput.csv", "hosts.csv"),
  VERBOSE = TRUE,
  callback = NULL
)
}
\arguments{
\item{flags}{Vector of boolean flags for simulation outputs. All of the output file names are specified in the output_files arugment.
1) save_SIR: if TRUE, saves the SIR dynamics
2) save_viruses: if TRUE, saves information on the simulation viruses.
3) use_time: if TRUE, records the duration of the simulation run
4) save_hosts: if TRUE, saves the properties of the simulation host population.}

\item{hostpars}{vector of parameters relating to the host population. In order, these are: S0, I0, R0, contact rate, birth/death rate, temporary immunity waning rate (duration of recovered period), true 0 viral load value, and the ID of the seed variant}

\item{vlPars}{pre-computed matrix of random viral load kinetics parameters}

\item{infectiousnessPars}{matrix of infectiousness parameters}

\item{crossImmunity}{matrix of cross immunity}

\item{start}{Simulation start day}

\item{end}{Simulation end day}

\item{output_files}{Vector of output file names. Note that these should be csv files. In order: 1) location of the SIR dynamics; 2) location of virus characteristics output; 3) Where to save the entire host population characteristics.}

\item{VERBOSE}{If TRUE, prints additional simulation output}

\item{callback}{Leave this - this is just used by the shiny app for the progress bar.}
}
\value{
Returns the ID of the last generated virus
}
\description{
Main simulation interface
}
