% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_simulation_cpp}
\alias{run_simulation_cpp}
\title{Interface to the entire C++ driftSim simulation}
\usage{
run_simulation_cpp(
  flags,
  hostPopn,
  vlPars,
  infectiousnessPars,
  crossImmunity,
  day,
  final_day,
  output_files,
  VERBOSE,
  callback
)
}
\arguments{
\item{flags}{set of boolean flags for simulation output. See \code{\link{run_simulation}}}

\item{hostPopn}{vector of parameters related to the host population. See \code{\link{run_simulation}}}

\item{vlPars}{Matrix of parameters related to the iviral load kinetics. See \code{\link{run_simulation}}}

\item{day}{integer of simulation starting day}

\item{final_day}{integer of simulation final day (steps of 1 day)}

\item{output_files}{vector of output file names, as in \code{\link{run_simulation}}}

\item{VERBOSE}{boolean to indicate how much console output the simulation should have}

\item{callback}{pointer to enable call backs to R. Pretty much just for progress bar in the shiny app}

\item{scenario}{integer 1 to 4, indicating which version of the model is being run. See \code{\link{scenario_descriptions}} for a description of each scenario}
}
\value{
the final virus ID from the simulation
}
\description{
Interface to the entire C++ driftSim simulation
}
