% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenic_drift.R
\name{scenario_descriptions}
\alias{scenario_descriptions}
\title{Descriptions of model scenarios}
\usage{
scenario_descriptions(scenario)
}
\arguments{
\item{scenario}{an integer 1-4 for the desired scenario}
}
\value{
the text description of the given scenario
}
\description{
Returns a text description of a given scenario 1-4
}

