% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setMigProbs}
\alias{setMigProbs}
\title{Set migration Probabilities}
\usage{
setMigProbs(n, mig)
}
\arguments{
\item{n}{The number of populations}

\item{mig}{The migration rate}
}
\value{
A matrix with migration probabilities
}
\description{
Defines the migration rate using a radial model with Pop1 as the central population.
All others form a ring aorund this population.
}
\examples{
setMigProbs(6, 0.1)
}

