% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDrift.R
\name{simDrift}
\alias{simDrift}
\title{Simulate Drift}
\usage{
simDrift(f0, N0, Nt, t, n, mig, surv, litter, ...)
}
\arguments{
\item{f0}{The allele frqeuency pre-bottleneck}

\item{N0}{The effective population size pre-bottleneck}

\item{Nt}{The effective population size after 't' generations}

\item{t}{The number of generations}

\item{n}{The number of populations}

\item{mig}{The migration rate}

\item{surv}{The survival rate at the original population bottleneck}

\item{litter}{The litter size at each generation}
}
\value{
A list with components \code{ft} and \code{nEff}.
These denote the final allele frequency, and the effective populations sizes at each geenration respectively.
}
\description{
Simulate drift across a number of populations and generations, for a single allele
}
\examples{
simDrift(f0 = 0.8, N0 = 100, Nt = 200, t = 10, n = 6, mig = 0.01, surv = 0.1, litter= 6)

}

