% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenic_drift.R
\name{generateHostKDist}
\alias{generateHostKDist}
\title{xx}
\usage{
generateHostKDist(hostFile, N)
}
\arguments{
\item{flags}{xx}

\item{hostpars}{xx}

\item{viruspars}{xx}

\item{start}{xx}

\item{end}{xx}

\item{output_files}{xx}

\item{VERBOSE}{xx}

\item{scenario}{xx}

\item{callback}{xx}
}
\description{
moar2
}

