% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenic_drift.R
\name{generateHostKDist_2}
\alias{generateHostKDist_2}
\title{Add this to the source file so that you can access it ie. cluster_submission.R}
\usage{
generateHostKDist_2(hostKs, N)
}
\description{
Add this to the source file so that you can access it ie. cluster_submission.R
}

