% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Simulate unbounded migration}
\usage{
migrate(pops, rate, method = "random")
}
\arguments{
\item{pops}{A list of populations with each containing a vector of individuals as 0, 2 or 2}

\item{rate}{The upper bound on the migration rate}

\item{method}{Does migration happen randomly between populations (method = "random")
or in a circular direction (method = "circular").}
}
\value{
A list of populations with the identical structure as supplied in pops.
The rate of shuffling between populations will have been performed according
to the migration rate and method specified in the function call.
}
\description{
Simulate migration between populations.
Equal migration between populations is assumed
The size of the first population is used to determine the number of migrants
}
\examples{
# Simulate 3 populations
pops <- lapply(c(10, 12, 14), function(x){sample(0:2, x, replace = TRUE)})
names(pops) <- paste0("Pop", LETTERS[seq_along(pops)])
# Add names so the individuals can be tracked
pops <- sapply(names(pops),
               function(x){
                 pop <- pops[[x]]
                 names(pop) <- paste(x, letters[seq_along(pop)], sep=":")
                 pop},
               simplify = FALSE)
# Now perform simulated migration
migrate(pops, 0.1, "random")
migrate(pops, 0.1, "circular")

}

