% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenic_drift.R
\name{run_simulation}
\alias{run_simulation}
\title{xx}
\usage{
run_simulation(flags = c(1, 0, 0, 0), hostpars = c(90000, 100, 1e+05 - 90000
  - 100, 1.5, 1/(40 * 365), 1/25, 0.333, 0.8), viruspars = c(3, 50, 1, 0.7, 3,
  2, 2, 0.1, 1, 0.5, 1000), start = 0, end = 365,
  output_files = c("SIR_output.csv", "voutput_1.csv", "voutput_2.csv"),
  VERBOSE = TRUE, scenario = 1)
}
\arguments{
\item{flags}{xx}

\item{hostpars}{xx}

\item{viruspars}{xx}

\item{start}{xx}

\item{end}{xx}

\item{output_files}{xx}

\item{VERBOSE}{xx}

\item{scenario}{xx}
}
\description{
moar
}

